refactor: update handlers' folder name so that it mentions "routes"
