+++ internal/api/routesHandlers/todos_routes.go
@@ -1,4 +1,4 @@
-package handlers
+package routeshandlers
 
 import (
 	"github.com/GuillaumeDeconinck/todos-go/internal/api/dao"
